Fix odd behaviour in splat HTML for the nested footnotes
Turned out that there were two nasty bugs around, where the HTML
footnotes where processed twice and not flushed correctly.

Added tests to check this.

Not really planned behaviour: in HTML the secondary footnotes which
are referenced from inside a primary footnote are appended at the end
of the document.

To keep the logical ordering, I think two stacks are needed.

Anyway, for now seems to work.

Address #24