Extend character set in directives.
Close #28

Directives are application-specific, as they provide values to the
application processing there. Now, the important thing in PR #28 is
that we need *parsing* compatibility with Emacs Muse, while the values
returned are relevant only in the application context.

So, permit underscores and dashes, but remove them from the values
stored and returned.