Reorganized anchor-related code
There are still bugs around, but it starts to look kind of ok.

The anchors are not parsed in the inline chunk, but at line parsing
time. This means that the anchor will point to a block, not to a
particular word (almost equivalent, as you usually point to something
to read, not to a word).

The adjacency logic is respected. Lonely anchors go to the next
element. If there are multiple anchors, the order is maintained.

Address #39