Follow up to URL formatting
Reference https://github.com/melmothx/text-amuse/pull/41

Preserve the monospacing font when formatting urls without labels.

If the users don't want that, they can always use the url with label,
which can be formatted freely.

This preserves the \url formatting while giving us the needed
flexibility.

Also add class to HTML so the class can be bound to a monospace font
as well.