Support page number reference interpolating '??'
When producing LaTeX output, interpolate the text inside a link to a
named anchor looking for a double question mark and interpolate it
with the page reference, so you can write [[#here][see page ??]]

The ?? placeholder is used because it's the same LaTeX output as an
undefined reference.

This of course leaves the HTML a bit crippled, but still functional.